+++ src/components/MyVocabulary/MyVocabulary.js
@@ -207,36 +207,40 @@ const MyVocabulary = (props) => {
 
   const removeWordsHandler = async () => {
     const wordsToRemove = props.visibleWords.filter(w => w.selected)
+    const lemmasToRemove = wordsToRemove.map(w => w.lemma)
     const idsToRemove = wordsToRemove.map(w => w.id)
-    try {
-      await learningdataService.removeMany(idsToRemove)
-      wordsToRemove.forEach(word => {props.removeWordFromMyVocabulary(word)})
-      if (wordsToRemove.length > 0) {
-        let lemmas = wordsToRemove[0].lemma
-        if (wordsToRemove.length > 1) {
-          wordsToRemove.shift()
-          lemmas = wordsToRemove.reduce((acc, cur) =>
-            `${acc}, ${cur.lemma}`, lemmas
-          )
-        }
-        const wordsRemoved = t('WordsRemovedMessage')
-        props.displayNotification({
-          message: `${wordsRemoved} ${lemmas}`,
-          messageType: 'success'
-        })
-        setNothingSelected(true)
-      }
-    } catch (error) {
-      console.log(error.response)
-      if (error.response.request.status === 404) {
+    let lemmas = lemmasToRemove[0]
+    if (lemmasToRemove.length > 1) {
+      lemmasToRemove.shift()
+      lemmas = lemmasToRemove.reduce((acc, cur) =>
+        `${acc}, ${cur}`, lemmas
+      )
+    }
+    const message = t('ConfirmRemove')
+    if (window.confirm(`${message} ${lemmas}?`)) {
+      try {
+        await learningdataService.removeMany(idsToRemove)
         wordsToRemove.forEach(word => {props.removeWordFromMyVocabulary(word)})
-        setNothingSelected(true)
-      } else {
-        props.displayNotification({
-          message: t('RemovingFailed'),
-          messageType: 'danger'
-        })
-        setNothingSelected(false)
+        if (wordsToRemove.length > 0) {
+          const wordsRemoved = t('WordsRemovedMessage')
+          props.displayNotification({
+            message: `${wordsRemoved} ${lemmas}`,
+            messageType: 'success'
+          })
+          setNothingSelected(true)
+        }
+      } catch (error) {
+        console.log(error.response)
+        if (error.response.request.status === 404) {
+          wordsToRemove.forEach(word => {props.removeWordFromMyVocabulary(word)})
+          setNothingSelected(true)
+        } else {
+          props.displayNotification({
+            message: t('RemovingFailed'),
+            messageType: 'danger'
+          })
+          setNothingSelected(false)
+        }
       }
     }
   }
